refactored Connection and added Line class
