+++ Code/src/main/java/nl/utwente/group10/ui/components/Connection.java
@@ -4,31 +4,25 @@ import java.io.IOException;
 
 import nl.utwente.cs.caes.tactile.fxml.TactileBuilderFactory;
 import nl.utwente.group10.ui.Main;
+import nl.utwente.group10.ui.components.Line;
 import javafx.scene.paint.Color;
-import javafx.scene.shape.Circle;
-import javafx.scene.shape.CubicCurve;
 import javafx.beans.property.DoubleProperty;
 import javafx.beans.value.ChangeListener;
 import javafx.beans.value.ObservableValue;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Node;
 import javafx.scene.control.Label;
-import javafx.scene.layout.Pane;
 import javafx.scene.layout.Region;
 import javafx.scene.layout.StackPane;
 import javafx.scene.control.Label;
-import javafx.scene.shape.Line;
+
 
 /**
  * This class represents a connection between two different FunctionBlocks. The
  * output of one FunctionBlock will be used as input for another FunctionBlock
  */
-public class Connection extends CubicCurve implements ChangeListener<Number> {
+public class Connection extends Line implements ChangeListener<Number> {
 
-	public static final double BEZIER_CONTROL_OFFSET_Y = 100f;
-	
-	private ConnectionAnchor startAnchor;
-	private ConnectionAnchor endAnchor;
 	/** The Block that inputs data into this connection */
 	private Block input;
 	/** The Block that we output data into from this connection */
@@ -51,28 +45,12 @@ public class Connection extends CubicCurve implements ChangeListener<Number> {
 		startAnchor.layoutYProperty().addListener(connection);
 		endAnchor.layoutXProperty().addListener(connection);
 		endAnchor.layoutYProperty().addListener(connection);
-		connection.startAnchor = startAnchor;
-		connection.endAnchor = endAnchor;			
-		connection.updateStartEndPositions();
+		
+		connection.setStartAnchor(startAnchor);
+		connection.setEndAnchor(endAnchor);	
 		
 		return connection;
 	}
-	
-	
-	public void setStartPosition(double x, double y) {
-		this.setStartX(x);
-		this.setStartY(y);
-		this.setControlX1(x);
-		this.setControlY1(y+BEZIER_CONTROL_OFFSET_Y);
-	}
-	
-	public void setEndPosition(double x, double y) {
-		this.setEndX(x);
-		this.setEndY(y);
-		this.setControlX2(x);
-		this.setControlY2(y-BEZIER_CONTROL_OFFSET_Y);
-	}
-	
 
 	/**
 	 * Private method that sets the input into this Connection
@@ -109,21 +87,7 @@ public class Connection extends CubicCurve implements ChangeListener<Number> {
 	}
 
 	@Override
-	public void changed(ObservableValue<? extends Number> observable,
-			Number oldValue, Number newValue) {
-		
+	public void changed(ObservableValue<?extends Number> observable, Number oldValue, Number newValue) {
 		updateStartEndPositions();
-	}
-	
-	private void updateStartEndPositions() {
-		updateStartPosition();
-		updateEndPosition();
-	}
-	
-	private void updateStartPosition() {
-		setStartPosition(startAnchor.getLayoutX()+startAnchor.getCenterX()-this.getLayoutX(),startAnchor.getLayoutY()+startAnchor.getCenterY()-this.getLayoutY());
-	}
-	private void updateEndPosition() {
-		setEndPosition(endAnchor.getLayoutX()+endAnchor.getCenterX()-this.getLayoutX(),endAnchor.getLayoutY()+endAnchor.getCenterY()-this.getLayoutY());
 	}
 }
